backend/requirements.txt
fastapi
uvicorn
PyPDF2
google-generativeai
reportlab
python-dotenv
python-multipart